+++ staircase/core/exceptions/__init__.py
@@ -3,10 +3,10 @@ General exceptions
 """
 
 
-class DifferentClosedValuesError(ValueError):
+class ClosedMismatchError(ValueError):
     def __init__(self, stairs1, stairs2):
         """
-        Create a `DifferentClosedValuesError` indicating that `stairs1` and
+        Create a `ClosedMismatchError` indicating that `stairs1` and
         `stairs2` were supposed to have the same `clased` values but didn't.
         """
         super().__init__(