+++ staircase/core/ops/arithmetic.py
@@ -6,10 +6,7 @@ import pandas as pd
 
 import staircase as sc
 from staircase.core.ops import docstrings
-from staircase.core.ops.common import (
-    _combine_stairs_via_values,
-    _requires_closeds_equal,
-)
+from staircase.core.ops.common import _combine_stairs_via_values, requires_closed_match
 from staircase.util import _sanitize_binary_operands
 from staircase.util._decorators import Appender
 
@@ -43,7 +40,7 @@ def _add_or_sub_deltas_no_mask(self, other, series_op, float_op):
 
 def _make_add_or_sub_func(docstring, series_op, float_op, series_rop):
     @Appender(docstring, join="\n", indents=1)
-    @_requires_closeds_equal
+    @requires_closed_match
     def func(self, other):
         self, other = _sanitize_binary_operands(self, other)
         if self._data is None and other._data is None:
@@ -108,7 +105,7 @@ subtract = _make_add_or_sub_func(
 
 def _make_mul_div_func(docstring, series_op, float_op, series_rop, float_rop):
     @Appender(docstring, join="\n", indents=1)
-    @_requires_closeds_equal
+    @requires_closed_match
     def func(self, other):
         def op_with_scalar(self, other, series_op, float_op):
             # other is scalar