+++ staircase/core/ops/logical.py
@@ -5,10 +5,7 @@ import pandas as pd
 
 import staircase as sc
 from staircase.core.ops import docstrings
-from staircase.core.ops.common import (
-    _combine_stairs_via_values,
-    _requires_closeds_equal,
-)
+from staircase.core.ops.common import _combine_stairs_via_values, requires_closed_match
 from staircase.util import _sanitize_binary_operands
 from staircase.util._decorators import Appender
 
@@ -78,7 +75,7 @@ def _make_logical_func(docstring, array_op, float_op):
     }[array_op]
 
     @Appender(docstring, join="\n", indents=1)
-    @_requires_closeds_equal
+    @requires_closed_match
     def func(self, other):
         self, other = _sanitize_binary_operands(self, other)
         if other._data is None: